requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

groq==0.11.0

langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0

sentence-transformers==3.1.0
torch==2.4.0

faiss-cpu==1.8.0
pinecone-client==4.1.0

numpy==1.26.4
pandas==2.2.2
httpx==0.27.2

structlog==24.4.0

streamlit==1.38.0
plotly==5.24.0